MRVL-6 [core-team] create HeroesDetails module and enrich models
