+++ Modules/HeroDetails/Sources/HeroDetails/HeroDetailsView.swift
@@ -1,0 +1,7 @@
+//
+//  HeroDetailsView.swift
+//  HeroDetails
+//
+//  Created by Omar Tarek Mansour Omar on 15/8/25.
+//
+
