+++ Modules/HeroesCore/Sources/HeroesCore/Data/DtoModels/HeroesResponseDto.swift
@@ -1,0 +1,11 @@
+//
+//  HeroesResponseDto.swift
+//  HeroesCore
+//
+//  Created by Omar Tarek Mansour Omar on 15/8/25.
+//
+import Foundation
+
+public struct HeroesResponseDto: Decodable {
+    let data: HeroesContainerDto
+}
