+++ Modules/HeroesCore/Sources/HeroesCore/Data/DtoModels/PaginationDto.swift
