+++ Modules/HeroesCore/Sources/HeroesCore/Data/DtoModels/ThumbnailDto.swift
@@ -1,0 +1,14 @@
+//
+//  ThumbnailDto.swift
+//  HeroesCore
+//
+//  Created by Omar Tarek Mansour Omar on 15/8/25.
+//
+
+import Foundation
+
+struct ThumbnailDto: Decodable {
+    let path: String
+    let `extension`: String
+}
+
