+++ Modules/HeroesCore/Sources/HeroesCore/Domain/Interactors/HeroesPaginationInteractor.swift
